+++ packages/expo-constants/src/__tests__/Constants-test.ts
@@ -12,6 +12,16 @@ it(`defines a linking URI and URL`, () => {
   expect(Constants.linkingUri).toBe(Constants.linkingUrl);
 });
 
+function mockExponentConstants(mockValues: object) {
+  jest.doMock('../ExponentConstants', () => {
+    const ExponentConstants = jest.requireActual('../ExponentConstants');
+    return {
+      ...ExponentConstants,
+      ...mockValues,
+    };
+  });
+}
+
 describe(`manifest`, () => {
   const fakeManifest = { id: '@jester/manifest' };
   const fakeManifest2 = { id: '@jester/manifest2' };
@@ -31,16 +41,6 @@ describe(`manifest`, () => {
     console.warn = jest.fn();
   });
   afterEach(() => (console.warn = originalWarn));
-
-  function mockExponentConstants(mockValues: object) {
-    jest.doMock('../ExponentConstants', () => {
-      const ExponentConstants = jest.requireActual('../ExponentConstants');
-      return {
-        ...ExponentConstants,
-        ...mockValues,
-      };
-    });
-  }
 
   function mockNativeModulesProxy(mockValues: object) {
     jest.doMock('@unimodules/core', () => {
@@ -74,14 +74,14 @@ describe(`manifest`, () => {
   it(`exists if defined as an object in ExponentConstants`, () => {
     mockExponentConstants({ manifest: fakeManifest });
     const ConstantsWithMock = require('../Constants').default;
-    expect(ConstantsWithMock.manifest).toEqual(fakeManifest);
+    expect(ConstantsWithMock.manifest).toMatchObject(fakeManifest);
     expect(console.warn).not.toHaveBeenCalled();
   });
 
   it(`exists if defined as a string in ExponentConstants`, () => {
     mockExponentConstants({ manifest: JSON.stringify(fakeManifest) });
     const ConstantsWithMock = require('../Constants').default;
-    expect(ConstantsWithMock.manifest).toEqual(fakeManifest);
+    expect(ConstantsWithMock.manifest).toMatchObject(fakeManifest);
     expect(console.warn).not.toHaveBeenCalled();
   });
 
@@ -89,7 +89,7 @@ describe(`manifest`, () => {
     mockExponentConstants({ manifest: undefined });
     mockExpoUpdates({ manifest: fakeManifest, manifestString: undefined });
     const ConstantsWithMock = require('../Constants').default;
-    expect(ConstantsWithMock.manifest).toEqual(fakeManifest);
+    expect(ConstantsWithMock.manifest).toMatchObject(fakeManifest);
     expect(console.warn).not.toHaveBeenCalled();
   });
 
@@ -97,7 +97,7 @@ describe(`manifest`, () => {
     mockExponentConstants({ manifest: undefined });
     mockExpoUpdates({ manifest: undefined, manifestString: JSON.stringify(fakeManifest) });
     const ConstantsWithMock = require('../Constants').default;
-    expect(ConstantsWithMock.manifest).toEqual(fakeManifest);
+    expect(ConstantsWithMock.manifest).toMatchObject(fakeManifest);
     expect(console.warn).not.toHaveBeenCalled();
   });
 
@@ -133,7 +133,7 @@ describe(`manifest`, () => {
     mockExponentConstants({ manifest: fakeManifest });
     mockExpoUpdates({ manifest: fakeManifest2 });
     const ConstantsWithMock = require('../Constants').default;
-    expect(ConstantsWithMock.manifest).toEqual(fakeManifest2);
+    expect(ConstantsWithMock.manifest).toMatchObject(fakeManifest2);
     expect(console.warn).not.toHaveBeenCalled();
   });
 
@@ -141,7 +141,7 @@ describe(`manifest`, () => {
     mockExponentConstants({ manifest: fakeManifest });
     mockExpoUpdates({ manifest: {} });
     const ConstantsWithMock = require('../Constants').default;
-    expect(ConstantsWithMock.manifest).toEqual(fakeManifest);
+    expect(ConstantsWithMock.manifest).toMatchObject(fakeManifest);
     expect(console.warn).not.toHaveBeenCalled();
   });
 
@@ -158,4 +158,35 @@ describe(`manifest`, () => {
       });
     });
   }
+});
+
+describe(`manifestDerivedMethods`, () => {
+  const legacyManifest = { id: '@jester/manifest', sdkVersion: '36.0.0' };
+  const newManifest = {
+    id: '6abaa89a-34e1-4c91-82fe-9139530dcf3f',
+    updateMetadata: {},
+    runtimeVersion: 'exposdk:36.0.0',
+  };
+
+  beforeEach(() => {
+    jest.resetModules();
+  });
+
+  afterEach(() => {
+    jest.dontMock('../ExponentConstants');
+  });
+
+  describe(`getSDKVersion`, () => {
+    it('returns SDK version for legacy manifest', () => {
+      mockExponentConstants({ manifest: legacyManifest });
+      const ConstantsWithMock = require('../Constants').default;
+      expect(ConstantsWithMock.manifestDerivedMethods.getSDKVersion()).toEqual('36.0.0');
+    });
+
+    it('returns SDK version for new manifest', () => {
+      mockExponentConstants({ manifest: newManifest });
+      const ConstantsWithMock = require('../Constants').default;
+      expect(ConstantsWithMock.manifestDerivedMethods.getSDKVersion()).toEqual('36.0.0');
+    });
+  });
 });